+++ packages/jsx/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Bump dependencies
+- Export webpack loader
 
 ## 0.1.2 (2022-09-18)